Ex_38

+++ Задача 38/Program.cs
@@ -3,15 +3,15 @@
 
 // [3 7 22 2 78] -> 76
 
-int[] array = new int [10];
+double[] array = new double [10];
 
 for (int i = 0; i < array.Length; i++)
 {
     array[i] = new Random(). Next(1, 10);
     Console.Write(array[i] + " , ");
 }
-int min = array[0];
-int max = array[0];
+double min = array[0];
+double max = array[0];
 for (int i = 0; i < array.Length; i++)
 {
     if (array[i] > max) max = array[i];
